+++ apps/frontend/src/app/5_pages/AavePage/components/BorrowAssetsList/BorrowAssetsList.types.tsx
